requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.0.2
python-dotenv==0.19.2
mysqlclient==2.1.0
gunicorn==20.0.4
flask_login==0.6.3
clerk_backend_api== 1.3.1  
psycopg2-binary==2.9.10